Crate backend

Hound Backend

The API and types for Hound’s backend service.

Modules

cli

Command Line Interface for the binary.

server

The HTTP server API.

trailbase_client

A client library to connect to a TrailBase server via HTTP.

Structs

AuthenticatedClient

A client that has been verified to authentication tokens, and contains all methods accessing the app backend that require authenticated access.

Client

The client for the API backend in the hound application. This type provides general utilities along with methods to access the backend. Note that this type only provides methods which are valid without authentication, to access authenicated methods you need to use AuthenticatedClient through Self::auth or Self::login.