Module backend::api::token

Structs

OAuthClientConfig

OAuth Public Config response struct

Response

simple response struct

TokenBody

TODO: look into this

give_auth_url

give oauth auth url in json message for http request

give_client_id

give oauth client id in json message for http request

give_config

give public parts of the oauth config in json format for http request

give_token_url

give oauth token url in json message for http request

login

This is a bad login do not use

token

TODO: what is this used for

Functions

config

config oauth api part of the actix

make_client 🔒

Helper to make a client that often used in this file