Enum backend::instructions::Instructions

pub enum Instructions {
    Clear,
    Return,
    Jump(Address),
    Call(Address),
    SkipEqualK(Vx, Byte),
    SkipNotEqualK(Vx, Byte),
    SkipEqual(Vx, Vy),
    StoreK(Vx, Byte),
    AddK(Vx, Byte),
    Store(Vx, Vy),
    Or(Vx, Vy),
    And(Vx, Vy),
    Xor(Vx, Vy),
    Add(Vx, Vy),
    Sub(Vx, Vy),
    ShiftRight(Vx, Vy),
    SubReversed(Vx, Vy),
    ShiftLeft(Vx, Vy),
    SkipNotEqual(Vx, Vy),
    StoreI(Address),
    JumpOffset(Address),
    Random(Vx, Byte),
    Draw(Vx, Vy, Nibble),
    SkipKeyPressed(Vx),
    SkipKeyNotPressed(Vx),
    StoreDelayTimer(Vx),
    WaitForKeypressAndStore(Vx),
    SetDelayTimer(Vx),
    SetSoundTimer(Vx),
    AddToI(Vx),
    LoadHexGlyph(Vx),
    StoreBCD(Vx),
    StoreRegisters(Vx),
    LoadRegisters(Vx),
    Invalid,
}

This contains all of the instructions in the Chip-8 instruction set.

Variants
Clear

Clear the screen.

Opcode: 00E0

Return

Return from a subroutine.

Opcode: 00EE

Jump(Address)

Jump to address NNN.

Opcode: 1NNN

Call(Address)

Call the subroutine at address NNN.

Opcode: 2NNN

SkipEqualK(Vx, Byte)

Skip the following instruction if the value of register VX equals NNN.

Opcode: 3XNN

SkipNotEqualK(Vx, Byte)

Skip the following instruction if the value of register VX does not equal NNN.

Opcode: 4XNN

SkipEqual(Vx, Vy)

Skip the following instruction if the value of register VX is equal to the value of register VY.

Opcode: 5XY0

StoreK(Vx, Byte)

Store number NNN in register VX.

Opcode: 6XNN

AddK(Vx, Byte)

Add the value NNN to register VX.

Opcode: 7XNN

Store(Vx, Vy)

Store the value of register VY in register VX.

Opcode: 8XY0

Or(Vx, Vy)

Set VX to VX OR VY.

Opcode: 8XY1

And(Vx, Vy)

Set VX to VX AND VY.

Opcode: 8XY2

Xor(Vx, Vy)

Set VX to VX XOR VY.

Opcode: 8XY3

Add(Vx, Vy)

Add the value of register VY to register VX.

Set VF to 01 if a carry occurs.

Set VF to 00 if a carry does not occur.

Opcode: 8XY4

Sub(Vx, Vy)

Subtract the value of register VY from register VX.

Set VF to 00 if a borrow occurs.

Set VF to 01 if a borrow does not occur.

Opcode: 8XY5

ShiftRight(Vx, Vy)

Store the value of register VY shifted right one bit in register VX.

Set register VF to the least significant bit prior to the shift.

Opcode: 8XY6

SubReversed(Vx, Vy)

Set register VX to the value of VY minus VX.

Set VF to 00 if a borrow occurs.

Set VF to 01 if a borrow does not occur.

Opcode: 8XY7

ShiftLeft(Vx, Vy)

Store the value of register VY shifted left one bit in register VX.

Set register VF to the most significant bit prior to the shift.

Opcode: 8XYE

SkipNotEqual(Vx, Vy)

Skip the following instruction if the value of register VX is not equal to the value of register VY.

Opcode: 9XY0

StoreI(Address)

Store memory address NNN in register I.

Opcode: ANNN

JumpOffset(Address)

Jump to address NNN + V0.

Opcode: BNNN

Random(Vx, Byte)

Set VX to a random number with a mask of NN.

Opcode: CXNN

Draw(Vx, Vy, Nibble)

Draw a sprite at position VX,VY with N bytes of sprite data starting at the address stored in I.

Set VF to 01 if any set pixels are changed to unset, and 00 otherwise.

Opcode: DXYN

SkipKeyPressed(Vx)

Skip the following instruction if the key corresponding to the hex value currently stored in register VX is pressed.

Opcode: EX9E

SkipKeyNotPressed(Vx)

Skip the following instruction if the key corresponding to the hex value currently stored in register VX is not pressed.

Opcode: EXA1

StoreDelayTimer(Vx)

Store the current value of the delay timer in register VX.

Opcode: FX07

WaitForKeypressAndStore(Vx)

Wait for a keypress and store the result in register VX.

Opcode: FX0A

SetDelayTimer(Vx)

Set the delay timer to the value of register VX.

Opcode: FX15

SetSoundTimer(Vx)

Set the sound timer to the value of register VX.

Opcode: FX18

AddToI(Vx)

Add the value stored in register VX to register I.

Opcode: FX1E

LoadHexGlyph(Vx)

Set I to the memory address of the sprite data corresponding to the hexadecimal digit stored in register VX.

Opcode: FX29

StoreBCD(Vx)

Store the binary-coded decimal equivalent of the value stored in register VX at addresses I, I+1, and I+2.

Opcode: FX33

StoreRegisters(Vx)

Store the values of register V0 to VX inclusive in memory starting at address I.

I is set to I + X + 1 after operation.

Opcode: FX55

LoadRegisters(Vx)

Fill registers V0 to VX inclusive with the values stored in memory starting at address I.

I is set to I + X + 1 after operation.

Opcode: FX65

Invalid

Any instruction received that does not belong to the Chip-8 instruction set.

Opcode: Unknown