backend::migrations::sea_orm::entity

Trait IdenStatic

pub trait IdenStatic:
    Iden
    + Copy
    + Debug
    + 'static {
    // Required method
    fn as_str(&self) -> &str;
}

Ensure the identifier for an Entity can be converted to a static str

Required Methods

fn as_str(&self) -> &str

Method to call to get the static string identity

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.