backend::migrations::sea_orm

Struct SelectThree

pub struct SelectThree<E, F, G>
where
    E: EntityTrait,
    F: EntityTrait,
    G: EntityTrait,
{ /* private fields */ }

Defines a structure to perform a SELECT operation on two Models

Implementations

impl<E, F, G, M, N, O> SelectThree<E, F, G>
where E: EntityTrait<Model = M>, F: EntityTrait<Model = N>, G: EntityTrait<Model = O>, M: FromQueryResult + Send + Sync, N: FromQueryResult + Send + Sync, O: FromQueryResult + Send + Sync,

pub fn cursor_by<C>(self, order_columns: C) -> Cursor<SelectThreeModel<M, N, O>>
where C: IdentityOf<E>,

Convert into a cursor using column of first entity

impl<E, F, G> SelectThree<E, F, G>
where E: EntityTrait, F: EntityTrait, G: EntityTrait,

pub fn into_model<M, N, O>(self) -> Selector<SelectThreeModel<M, N, O>>
where M: FromQueryResult, N: FromQueryResult, O: FromQueryResult,

Perform a conversion into a SelectThreeModel

pub fn into_partial_model<M, N, O>(self) -> Selector<SelectThreeModel<M, N, O>>
where M: PartialModelTrait, N: PartialModelTrait, O: PartialModelTrait,

Perform a conversion into a SelectThreeModel with PartialModel

pub fn into_json(self) -> Selector<SelectThreeModel<Value, Value, Value>>

Convert the Models into JsonValue

pub async fn one<C>( self, db: &C, ) -> Result<Option<(<E as EntityTrait>::Model, Option<<F as EntityTrait>::Model>, Option<<G as EntityTrait>::Model>)>, DbErr>
where C: ConnectionTrait,

Get one Model from the Select query

pub async fn all<C>( self, db: &C, ) -> Result<Vec<(<E as EntityTrait>::Model, Option<<F as EntityTrait>::Model>, Option<<G as EntityTrait>::Model>)>, DbErr>
where C: ConnectionTrait,

Get all Models from the Select query

pub async fn stream<'a, 'b, C>( self, db: &'a C, ) -> Result<impl Stream<Item = Result<(<E as EntityTrait>::Model, Option<<F as EntityTrait>::Model>, Option<<G as EntityTrait>::Model>), DbErr>> + 'b, DbErr>
where 'a: 'b, C: ConnectionTrait + StreamTrait + Send,

Stream the results of a Select operation on a Model

pub async fn stream_partial_model<'a, 'b, C, M, N, O>( self, db: &'a C, ) -> Result<impl Stream<Item = Result<(M, Option<N>, Option<O>), DbErr>> + Send + 'b, DbErr>
where 'a: 'b, C: ConnectionTrait + StreamTrait + Send, M: PartialModelTrait + Send + 'b, N: PartialModelTrait + Send + 'b, O: PartialModelTrait + Send + 'b,

Stream the result of the operation with PartialModel