Module backend::pool

Module for Database Pool

pool is a collection of Database Connection Pools that can be used to perform operations on multiple databases, e.g. PostgreSQL, Redis, Pulsar, Typesense, Minio, etc.

Structs

MinioImagePool

MinioImageStorage

Minio Connection Pool

PgDb

PostgreSQL Connection Pool

PulsarProducerPool

PulsarSearchProducerMq

Pulsar Connection Pool

RedisDb

Redis Connection Pool

RedisPoolWrapper

SeaOrmPool

SearchClient

TypesenseSearch

Typesense Connection Pool

TypesenseSearchPool

Traits

DeadManager

Manager used for managing redis connection pool.

Search