backend::server::webrtc

Function handle_signaling

async fn handle_signaling(
    ws: WebSocket,
    clients: Arc<Mutex<HashMap<i64, SplitSink<WebSocket, Message>>>>,
)

Receive and respond to messages on the “/signaling” end point Messages are json objects with some key “type”:“foo”, where foo is

join - the initial message, establishing the connection between the client and front-end, generating the client id, and adding the connection to the clients map.

candidate - iterate through the connected clients and send the candidate string from this message to those clients whose client id do not match this client’s

offer - iterate through the connected clients and send an offer string from this message to those clients whose client id does not match this client’s

answer - send an answer string to the client identified by the client id contained within this message